Relocate Gauss/shape modules from utilities/ to format/
The architecture proposal called for an MPCO format-policy package
that holds every per-element-class quirk in one place. The shape
that landed has format_policy in io/, but the Gauss-point catalog
and shape functions stayed under utilities/ — leaving "format
quirks" split across two trees. This PR consolidates them.

Layout
------
- New: src/STKO_to_python/format/__init__.py — re-exports the public
  API for one-import access.
- Move: utilities/gauss_points.py    -> format/gauss_points.py
- Move: utilities/shape_functions.py -> format/shape_functions.py
- Old paths become PEP 562 __getattr__ shims emitting
  DeprecationWarning per attribute access.

Decisions
---------
The proposal also envisioned a wrapper class (GaussPointMapper) and
an optional `coords="natural" | "global"` flag on element fetches.
Both are deliberately *not* introduced here:
- The registries (ELEMENT_IP_CATALOG, SHAPE_FUNCTIONS) and free
  functions (compute_physical_coords, compute_jacobian_dets) are
  simpler than a class wrapper and have no per-instance state to
  carry. The dispatch already happens via dict lookup.
- The "coords flag changing return shape" idea is an anti-pattern;
  the current ElementResults.physical_coords() / .jacobian_dets()
  separate-method API is cleaner and stays.

These two decisions are decided here and will be reflected in the
upcoming architecture docs page (Phase 5b).

Internal updates
----------------
- elements/elements.py and elements/element_results.py now import
  from format/ directly (no DeprecationWarning trip).

Tests
-----
- The three test files that hit the old utilities/ path are
  switched to format/. The deprecation parametrize block in
  test_public_api.py grows four cases (one per old-path symbol)
  to lock in the warning + same-object-identity contract.
- Full suite: 620 passed.
